requirements.txt
# Install with 'pip install -r requirements.txt'

# Infrastructure
Django >= 1.8, < 1.9
djangorestframework == 3.2
django-filter >= 0.9.2
markdown >= 2.1.0